feat: fix some condition and separate skills & projects in another file
