+++ src/data/projects.tsx
@@ -1,0 +1,115 @@
+export const projects = [
+  {
+    title: "SIAPPMI",
+    description:
+      "An information system for managing and tracking overseas worker deployment, covering registration, document processing, and status monitoring.",
+    image: "./img/siappmi.png",
+    technologies: [
+      "PHP",
+      "Laravel",
+      "MySQL",
+      "tFPDF",
+      "Bootstrap CSS",
+      "jQuery",
+    ],
+    demo: false,
+    openSource: false,
+  },
+  {
+    title: "Palomade",
+    description:
+      "An Palm fruit sorting app that helps farmers and processors identify ripe fruit, reduce waste, and track palm oil shipments efficiently.",
+    image: "/img/palomade.png",
+    technologies: [
+      "Node.js",
+      "Express",
+      "MySQL",
+      "Sequelize",
+      "Google Maps API",
+      "Cloud Run",
+      "Python",
+      "Tensorflow",
+      "Flask",
+      "Jetpack Compose",
+      "Docker",
+    ],
+    demo: false,
+    openSource: true,
+    repoUrl: "https://github.com/CH2-PS324",
+  },
+  {
+    title: "Commodity Price Predict",
+    description:
+      "A website dashboard serving a TensorFlow model to predict commodity prices, developed as part of my college final project.",
+    image: "./img/commodity-price-predict.png",
+    technologies: ["Python", "Flask", "Jinja2", "Tensorflow", "Bootstrap CSS"],
+    demo: false,
+    openSource: true,
+    repoUrl: "https://github.com/hafizcode02/comodity-price-predict",
+  },
+  {
+    title: "LPK BKJ Website",
+    description:
+      "A company profile website for a Japan job training institution.",
+    image: "./img/lpk-bkj.png",
+    technologies: ["PHP", "Laravel", "MySQL", "Bootstrap CSS", "jQuery"],
+    demo: true,
+    openSource: false,
+    demoUrl: "https://www.lpkbkj.co.id/",
+  },
+  {
+    title: "Baznas Kuningan Website",
+    description:
+      "A web app for donations, infaq, zakat payments, and information about BAZNAS Kuningan.",
+    image: "./img/baznas-kuningan.png",
+    technologies: [
+      "PHP",
+      "Laravel",
+      "MySQL",
+      "Bootstrap CSS",
+      "jQuery",
+      "Midtrans API",
+    ],
+    demo: false,
+    openSource: false,
+  },
+  {
+    title: "SIM-SURAT",
+    description:
+      "A web-based application for managing and tracking incoming and outgoing mail, including document scanning and archiving.",
+    image: "./img/sim-surat.png",
+    technologies: ["PHP", "Laravel", "Boostrap CSS", "MySQL"],
+    demo: false,
+    openSource: true,
+    repoUrl: "https://github.com/hafizcode02/sim-surat",
+  },
+  {
+    title: "Kejawanan Beach Website",
+    description:
+      "An informational website about Kejawanan Beach, Cirebon, Indonesia, created for the HIMIT PENS Event 2022.",
+    image: "./img/kejawanan.png",
+    technologies: ["HTML", "CSS", "Bootstrap CSS", "JavaScript", "Figma"],
+    demo: true,
+    openSource: true,
+    demoUrl: "https://kejawanan.netlify.app/",
+    repoUrl: "https://github.com/dimasrizkym/kejawanan",
+  },
+  {
+    title: "Carougen",
+    description:
+      "A content generator tool for creating and customizing carousels for instagram and other social media platforms.",
+    image: "./img/carougen.png",
+    technologies: ["Python", "Flask", "Jinja2", "OpenAI API", "Tailwind CSS"],
+    demo: false,
+    openSource: false,
+  },
+  {
+    title: "SIAKAD INVADA",
+    description:
+      "An Academic Information System for managing student data, course schedules, course enrollment, and exam results.",
+    image: "./img/siakad.png",
+    technologies: ["PHP", "Codeigniter 3", "Boostrap CSS", "MySQL"],
+    demo: false,
+    openSource: false,
+  },
+];
